+++ src/NRediSearch/QueryBuilder/UnionNode.cs
@@ -1,9 +1,0 @@
-﻿// .NET port of https://github.com/RedisLabs/JRediSearch/
-
-namespace NRediSearch.QueryBuilder
-{
-    public class UnionNode : QueryNode
-    {
-        protected override string GetJoinString() => "|";
-    }
-}